requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
bcrypt==4.0.1
